Revert "Revert "Ledger API Server refactor (#1077)" (#1096)"
This reverts commit e2fa13e62b09aa803ddfc245c01d1fb55a425f63.
