+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/sandbox/SandboxApplication.scala
@@ -55,9 +55,6 @@ object SandboxApplication {
       () => materializer.executionContext,
       () => {
         stopHeartbeats()
-        server.closeAllServices()
-      },
-      () => {
         server.close() // fully tear down the old server.
         buildAndStartServer(SqlStartMode.AlwaysReset)
       },
